Split benched.rs into separate Neo4j and Doublets modules
Addresses user feedback requesting clear separation of Neo4j and Doublets
logic. The benched.rs file contained mixed implementations for both databases.

Changes:
- Removed rust/src/benched.rs (contained both Neo4j and Doublets impls)
- Added rust/src/benched/mod.rs with common Benched trait
- Added rust/src/benched/doublets_benched.rs with Doublets implementations
- Added rust/src/benched/neo4j_benched.rs with Neo4j implementations

This provides complete separation between Neo4j and Doublets code across
all source files for easier side-by-side comparison.
